Refactor directory layout
