+++ app/components/ActivityCard.tsx
