+++ app/components/JournalEntryCard.tsx
