+++ app/components/MedicalRecordCard.tsx
