+++ app/components/PhotoGallery.tsx
